refactor(masks): hoist shared autocast and device helpers
- Move `_autocast_context` out of `core/masks/cutie.py` and `core/masks/sam.py` into `core/masks/base.py`; the two copies were byte-identical apart from a Cutie-specific note in the docstring, which the hoisted version keeps
- Add `_validate_device` to `utils/device.py`, beside `_best_device`; it validates a device the caller requested and raises `RuntimeError` when CUDA is asked for but `torch.cuda.is_available()` is False, so a typo or a CPU-only install surfaces before model weights load
- Document why `_validate_device` and `_best_device` stay separate: one validates a request, the other selects a device

---
